Created qa_single_streamer_loopback python unit test for crimson
Overview
=-=-=-=-

qa_single_streamer_loopback is the first unit test (technically functional or integration test) to use the new abstract test classes defined in crimson_qa.py.

The hope is to collect a large amount of common functional test code in crimson_qa.py and have an assortment of common functionality available for all classes that derive from those in crimson_qa.py .

For example, in crimson_qa.py, there are currently 4 abstract classes:

1) base
  - generic code applicable to all test configurations
  - sane values for all channels w.r.t. sample rate, gain, etc
2) straight_loopback
  - fills in some functionality given that there will always be pairs of channels operating at the same frequency
  - e.g. ( TX A, RX A ), ( TX C, RX C )
3) - multi_streamer_lb
  - derives from straight loopback but generates a different pair of streamers for each channel
4) - single_streamer_lb
  - derives from straight loopback and generates one pair of streamers for all channels

Current Issues
=-=-=-=-=-=-=-

We have one particular test that fails that compares the TX and RX signals and calls assertComplexTuplesAlmostEqual2().

There is a known issue with the FPGA at SoB. Either the RX start time is incorrectly reported or there is a variable delay in the CIC filter that we need to account for when packetizing, but often there is
a bunch of noise on RX until the correct signal begins to come through. At steady state, it is not an issue. However, this is a fairly critical building block for all test that we will ever perform. Indeed, it
is an assumption at this point that TX SoB and RX SoB happen and can be captured accurately.

Directions
=-=-=-=-=-

There is still the whole shiptest to run
* It would be helpful to implement short-time fourier transform (e.g. a sliding-window in time)
* We need algorithms to automatically pick out / count tones, harmonics, and spurs over the life of the transmission

It would be very helpful to have a collection of QA blocks in GNU Radio for the above, if things can be done in real-time. For analysis that is potentially too intensive to do in real-time, we need a collection of algorithms to do it offline.

* More advanced performance testing should be don in C++ (e.g. a test to stream all 4 channels at once, count dropped rx packets, etc)

I would definitely like to look into using Cheetah, which is (I believe) a template engine for python. Doing so would allow us to parametrise our testing.

I would also like to generate images corresponding to the correct number of channels per test so that we can create a nice-looking report.