stashing rx phase test
No time to work on QA any longer. Need to work on bringup of stratix 10, so I'm just checking in what I wrote.

Observations:

It the rx phase test does not seem to work consistently, and from what it appears, rx stream commands after the first one do not seem to have an effect. This is a bit troublesome. It would have been great to get rid of this issue before bringing up our next product, but we are running out of time. There is still so much work to be done, but I hope to reuse as much code as possible to hopefully halve the amount of potential work to do.